requirements.txt
# Core LangChain ecosystem
langchain==0.3.24
langchain-core==0.3.56
langchain-community==0.3.22
langchain-text-splitters==0.3.8
langgraph==0.3.34

# Google Gemini integration
google-genai
langchain-google-genai

# Hugging Face integration (fixed)
langchain-huggingface==0.1.2

# Hugging Face Hub & embeddings
huggingface_hub==0.30.2
sentence-transformers==4.1.0
faiss-cpu==1.10.0
chromadb==1.0.7

# PDF handling
pypdf==5.4.0

# Environment & tokenization
python-dotenv==1.1.0
tiktoken==0.9.0
pydantic==2.11.1

# Databases & notebooks
sqlalchemy==2.0.40
jupyter==1.1.1
